backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["dhilipsiva <dhilipsiva@pm.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.8"
graphene-django = "^2.12.1"
django-cors-headers = "^3.4.0"
django-extensions = "^3.0.3"
psycopg2 = "^2.8.5"
orm-choices = "^1.0.0"
whitenoise = "^5.1.0"
celery = "^4.4.6"
django-graphql-jwt = "^0.3.1"
boto3 = "^1.14.28"
eventlet = "^0.25.2"
python-socketio = "^4.6.0"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]
neovim = "^0.3.1"
ipython = "^7.16.1"
ipdb = "^0.13.3"
flake8 = "^3.8.3"
isort = "^5.1.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
